SSCS-11614 overloads issueDocument
