+++ src/main/java/uk/gov/hmcts/reform/sscs/util/PdfRequestUtil.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.reform.sscs.util;
 
-import static java.util.Objects.nonNull;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 
 import java.time.LocalDate;
@@ -8,7 +7,6 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import uk.gov.hmcts.reform.sscs.ccd.callback.DocumentType;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.sscs.ccd.domain.*;
@@ -23,8 +21,9 @@ public class PdfRequestUtil {
     private static String title;
     private static StringBuilder additionalRequestDetails;
 
-    @Value("${feature.postHearings.enabled}")
-    private static boolean isPostHearingsEnabled;
+    private PdfRequestUtil() {
+        //
+    }
 
     @AllArgsConstructor
     public enum PdfType {
@@ -54,12 +53,10 @@ public class PdfRequestUtil {
         StringBuilder pdfUrlBuilder = new StringBuilder();
         switch (pdfType) {
             case POST_HEARING:
-                if (isPostHearingsEnabled) {
-                    String postHearingDocumentTypeLabel = getPostHearingDocumentType(sscsCaseData).getLabel();
-                    pdfUrlBuilder.append(postHearingDocumentTypeLabel)
-                        .append(" from FTA");
-                    handlePostHearing(sscsCaseData);
-                }
+                String postHearingDocumentTypeLabel = getPostHearingDocumentType(sscsCaseData).getLabel();
+                pdfUrlBuilder.append(postHearingDocumentTypeLabel)
+                    .append(" from FTA");
+                handlePostHearing(sscsCaseData);
                 break;
             case POSTPONEMENT:
                 pdfUrlBuilder.append("Postponement Request");
@@ -188,17 +185,6 @@ public class PdfRequestUtil {
                 throw new IllegalArgumentException("Unexpected request type: " + postHearing.getRequestType());
         }
         return documentType;
-    }
-
-    public static String getEmbeddedDocumentTypeLabelForPostHearing(SscsCaseData caseData, String embeddedDocumentTypeLabel) {
-        PostHearing postHearing = caseData.getPostHearing();
-        if (isPostHearingsEnabled && nonNull(postHearing.getSetAside().getAction())) {
-            CcdCallbackMap action = postHearing.getSetAside().getAction();
-            boolean isGrantOrRefuseSetAsideAction = action.toString().equals(SetAsideActions.GRANT.getCcdDefinition())
-                    || action.toString().equals(SetAsideActions.REFUSE.getCcdDefinition());
-            embeddedDocumentTypeLabel =  isGrantOrRefuseSetAsideAction ? "Set Aside Decision Notice" : embeddedDocumentTypeLabel;
-        }
-        return embeddedDocumentTypeLabel;
     }
 
 }
